Refactored components: - marzipanopage, scene, hotspotcontainer; - usemarzipano hook; - and created useHotspotManager hook
